Fix client transaction deserialization issue
