+++ mempool/Cargo.toml
@@ -8,6 +8,7 @@ publish = false
 [dependencies]
 async-trait = "0.1.42"
 tokio = { version = "0.3.1", features = ["rt", "net", "time", "macros", "sync"] }
+tokio-util = { version = "0.5", features= ["codec"] }
 ed25519-dalek = "1.0.1"
 thiserror = "1.0.21"
 serde = { version = "1.0", features = ["derive"] }
